integrate network with protocol
