+++ Cargo.toml
@@ -21,7 +21,6 @@ members = [
     "core/wasm",
     "core/wasm/runtest",
     "core/wasm/runtest/generate-wasm/import-memory",
-    "core/near-network",
     "node/beacon",
     "node/client",
     "node/configs",
