--- core/near-network/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "near-network"
-version = "0.1.0"
-authors = ["Near Protocol <hello@nearprotocol.com>"]
-edition = "2018"
-
-[dependencies]
-tokio = "0.1"
-tokio-codec = "0.1"
-futures = "0.1"
-bytes = "0.4"
-log = "0.4"
-serde = "1.0"
-serde_derive = "1.0"
-rand = "0.6"
-parking_lot = "0.7"
-tokio-serde-cbor = "0.3"
-
-primitives = { path = "../../core/primitives"}
